Update README.md
the original instruction about using the interpreter were unclear and confusing to say the least. Why not just open a CMD window and install the requirements while transferring the files in the first place. Installing requirements via IDA will only work 50% of the time and is really a method for advanced users anyways. 90% of the time the IDA python script handler will catch on a new line and fail to resolve the script.

Installing the requirements.txt immediately after placing the file where it belongs is the proper way of doing things. You''ll then open IDA and have a properly working Gepetto plugin , the other method requires restarting IDA.