Improve translation installation 🎉✨
- Added `translate` as a global variable to improve visibility in the `load_config` function, as it is now essential to hold the translation reference. 🗣️

There's actually a very subtle bug in Gepetto. A classic problem with
`gettext`'s default behavior in environments where we do not fully control all
the code running in the same interpreter, like an embedded Python interpeter in
IDA, is that the `_` collides with any plugin or code that uses the `_` in
any other piece of code -- a very common pattern across Python programs. The
issue stems from `gettext.install()` sticking `_` directly
into Python's global namespace. So, if another plugin or piece of code (like in
IDA's embedded Python) uses _, it can accidentally overwrite ours.

Secondly, the
[documentation](https://docs.python.org/3/library/gettext.html#gettext.NullTranslations.install)
says that the `names` parameter (`install(names=None)`) must be given a
`sequence` -- and Python has a wonderful subtle unintuitive behavior where
strings are sequences. The [`install(names=None)`](https://github.com/python/cpython/blob/3.13/Lib/gettext.py#L329) method expects a list or
tuple. Passing a string like `translate.install("gepetto")` mistakenly sets `_`
for each letter `{'g', 'e', 'p', 't', 'o'}`, which definitely isn't what we want.

There are a few ways to handle it, I've started it here but wanted to raise it
as a point of discussion. A couple of ways we might handle this:

- Using a distinct global like `_gtr` to avoid any naming clashes.
- Using the translation object directly and skipping global installation altogether.

Open to feedback on the best approach here!